+++ cloud/docker/dockerfiles/Dockerfile-teagent
@@ -53,10 +53,10 @@ RUN apt-get update -y \
  && curl -fsSL https://download.docker.com/linux/ubuntu/gpg |  apt-key add - \
  && add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/ubuntu  $(lsb_release -cs) stable" \
  && apt-get update -y  \ 
- && apt-get install --no-install-recommends docker-ce jq -y \
+ && apt-get install --no-install-recommends docker-ce -y \
+ && apt-get install --install-suggests jq -y \
  && curl -L git.io/weave -o /usr/local/bin/weave \
  && chmod a+x /usr/local/bin/weave \
- && apt-get remove curl -y \
  && rm -rf /var/lib/apt/lists/* 
  
 ARG BE_PRODUCT_VERSION
