+++ cloud/docker/lib/run
@@ -18,19 +18,6 @@ fi
 
 BE_HOME=/opt/tibco/be/$SHORT_VERSION
 
-COMPONENT_PARSED="${COMPONENT%\"}"
-COMPONENT_PARSED="${COMPONENT_PARSED#\"}"
-
-echo "COMPONENT: $COMPONENT_PARSED"
-
-if [ "$COMPONENT_PARSED" = "tea" ]; then
-	TRA_FILE=$BE_HOME/teagent/bin/be-teagent.tra
-elif [ "$COMPONENT_PARSED" = "rms" ]; then
-	TRA_FILE=$BE_HOME/rms/bin/be-rms.tra
-else
-	TRA_FILE=$BE_HOME/bin/be-engine.tra
-fi
-
 if [[ -f "$BE_HOME/bin/be-engine.tra.original" ]]; then
 	rm -f $BE_HOME/bin/be-engine.tra
 	cp $BE_HOME/bin/be-engine.tra.original $BE_HOME/bin/be-engine.tra
@@ -42,6 +29,8 @@ CURRENT_DIR=$(pwd)
 BE_PROPS_FILE=beprops_all.props
 APP_HOME=/opt/tibco/be/application
 
+TRA_FILE=$BE_HOME/bin/be-engine.tra
+
 export PERL5LIB="/home/tibco/be"
 MAKE_BE_PROPS_RESULT=$(perl -Mbe_docker_run -e "be_docker_run::makeBeProps('$BE_PROPS_FILE','$APP_HOME','$BE_HOME/')")
 
@@ -51,6 +40,9 @@ if [[ -f "/home/tibco/be/gvproviders/run.sh" ]]; then
 fi
 
 /bin/cat $BE_PROPS_FILE
+
+COMPONENT_PARSED="${COMPONENT%\"}"
+COMPONENT_PARSED="${COMPONENT_PARSED#\"}"
 
 JMX_PORT_PARSED="${JMX_PORT%\"}"
 JMX_PORT_PARSED="${JMX_PORT_PARSED#\"}"
@@ -74,6 +66,7 @@ if [ "$COMPONENT_PARSED" = "tea" ]; then
   BE_TEA_AGENT_MONITORING_ONLY_PARSED="${BE_TEA_AGENT_MONITORING_ONLY_PARSED#\"}"
 
   TEA_PROPS_FILE="$BE_HOME/teagent/config/be-teagent.props"
+  TRA_FILE="$BE_HOME/teagent/bin/be-teagent.tra"
 
   echo "" >> $TEA_PROPS_FILE
   echo "be.tea.agent.monitoring.application=$BE_TEA_AGENT_MONITORING_ONLY_PARSED" >> $TEA_PROPS_FILE
@@ -95,63 +88,42 @@ if [ "$COMPONENT_PARSED" = "tea" ]; then
 
   BE_INSTANCE_POLLAR_INTERVAL_PARSED="${BE_INSTANCE_POLLAR_INTERVAL%}"
   BE_INSTANCE_POLLAR_INTERVAL_PARSED="${BE_INSTANCE_POLLAR_INTERVAL_PARSED#\"}"
-
+  
   #If discovery is for aws
-  if [[ "$BE_INSTANCE_DISCOVERY_TYPE_PARSED" == "aws" ]]; then
-	STACK_NAME_PARSED="beappstack"
-	STACK_NAME_PARSED="${STACK_NAME%}"
-    STACK_NAME_PARSED="${STACK_NAME_PARSED#\"}"
-    
-	AWS_ACCESS_KEY_PARSED="${AWS_ACCESS_KEY%}"
-	AWS_ACCESS_KEY_PARSED="${AWS_ACCESS_KEY_PARSED#\"}"
+  if [[ "$AWS_EXECUTION_ENV" == "AWS_ECS_EC2" ]] || [[ "$AWS_EXECUTION_ENV" == "AWS_ECS_FARGATE" ]]; then #If discovery is for aws fargate
 	
-	AWS_ACCESS_SECRET_PARSED="${AWS_ACCESS_SECRET%}"
-	AWS_ACCESS_SECRET_PARSED="${AWS_ACCESS_SECRET_PARSED#\"}"
+	if [ "$CLUSTER_NAME" != "" ]; then
+	  CLUSTER_NAME_PARSED="${CLUSTER_NAME%}"
+	  CLUSTER_NAME_PARSED="${CLUSTER_NAME_PARSED#\"}"
+	else
+	  CLUSTER_NAME_PARSED="nil"
+	fi
 	
-	AWS_REGION_PARSED="${AWS_REGION%}"
-	AWS_REGION_PARSED="${AWS_REGION_PARSED#\"}"
-	
-	AWS_OUTPUT_FORMAT_PARSED="${AWS_OUTPUT_FORMAT%}"
-	AWS_OUTPUT_FORMAT_PARSED="${AWS_OUTPUT_FORMAT_PARSED#\"}"
-	
-	AWS_ROLE_ARN_PARSED="${AWS_ROLE_ARN%}"
-	AWS_ROLE_ARN_PARSED="${AWS_ROLE_ARN_PARSED#\"}"
- 	
-	RUNCMD="${RUNCMD} $TEA_SERVER_URL_PARSED $TEA_SERVER_USERNAME_PARSED  $TEA_SERVER_PASSWORD_PARSED $PYTHONPATH $BE_INSTANCE_POLLAR_INTERVAL_PARSED  $STACK_NAME_PARSED $BE_INSTANCE_DISCOVERY_TYPE_PARSED $AWS_ACCESS_KEY_PARSED $AWS_ACCESS_SECRET_PARSED $AWS_REGION_PARSED $AWS_OUTPUT_FORMAT_PARSED $AWS_ROLE_ARN_PARSED"
-  #If discovery is for aws fargate
-  elif [[ "$BE_INSTANCE_DISCOVERY_TYPE_PARSED" == "fargate" ]]; then
-	CLUSTER_NAME_PARSED="${CLUSTER_NAME%}"
-	CLUSTER_NAME_PARSED="${CLUSTER_NAME_PARSED#\"}"
-	
-	BE_TEA_AGENT_URL_PARSED="${BE_TEA_AGENT_URL%}"
-	BE_TEA_AGENT_URL_PARSED="${BE_TEA_AGENT_URL_PARSED#\"}"
-	
-	AWS_ACCESS_KEY_PARSED="${AWS_ACCESS_KEY%}"
-	AWS_ACCESS_KEY_PARSED="${AWS_ACCESS_KEY_PARSED#\"}"
-	
-	AWS_ACCESS_SECRET_PARSED="${AWS_ACCESS_SECRET%}"
-	AWS_ACCESS_SECRET_PARSED="${AWS_ACCESS_SECRET_PARSED#\"}"
-	
-	AWS_REGION_PARSED="${AWS_REGION%}"
-	AWS_REGION_PARSED="${AWS_REGION_PARSED#\"}"
-	
-	AWS_OUTPUT_FORMAT_PARSED="${AWS_OUTPUT_FORMAT%}"
-	AWS_OUTPUT_FORMAT_PARSED="${AWS_OUTPUT_FORMAT_PARSED#\"}"
-	
-	AWS_ROLE_ARN_PARSED="${AWS_ROLE_ARN%}"
-	AWS_ROLE_ARN_PARSED="${AWS_ROLE_ARN_PARSED#\"}"
-	
-	RUNCMD="${RUNCMD} $TEA_SERVER_URL_PARSED $TEA_SERVER_USERNAME_PARSED  $TEA_SERVER_PASSWORD_PARSED $PYTHONPATH $BE_INSTANCE_POLLAR_INTERVAL_PARSED  $CLUSTER_NAME_PARSED $BE_INSTANCE_DISCOVERY_TYPE_PARSED $AWS_ACCESS_KEY_PARSED $AWS_ACCESS_SECRET_PARSED $AWS_REGION_PARSED $AWS_OUTPUT_FORMAT_PARSED $AWS_ROLE_ARN_PARSED"
-  #If discovery is for k8s
-  elif [[ "$BE_INSTANCE_DISCOVERY_TYPE_PARSED" == "k8s" ]]; then
-	BE_TEA_AGENT_URL_PARSED="${BE_TEA_AGENT_URL%}"
-    BE_TEA_AGENT_URL_PARSED="${BE_TEA_AGENT_URL_PARSED#\"}"
-    
-	RUNCMD="${RUNCMD}  $TEA_SERVER_URL_PARSED $TEA_SERVER_USERNAME_PARSED  $TEA_SERVER_PASSWORD_PARSED $PYTHONPATH $BE_INSTANCE_POLLAR_INTERVAL_PARSED $BE_TEA_AGENT_URL_PARSED $BE_INSTANCE_DISCOVERY_TYPE_PARSED"
+	if [ "$AWS_REGION" != "" ]; then
+	  AWS_REGION_PARSED="${AWS_REGION%}"
+	  AWS_REGION_PARSED="${AWS_REGION_PARSED#\"}"
+	else
+	  AWS_REGION_PARSED="nil"	  
+	fi
+
+	if [ "$AWS_OUTPUT_FORMAT" != "" ]; then
+	  AWS_OUTPUT_FORMAT_PARSED="${AWS_OUTPUT_FORMAT%}"
+	  AWS_OUTPUT_FORMAT_PARSED="${AWS_OUTPUT_FORMAT_PARSED#\"}"
+	else
+	  AWS_OUTPUT_FORMAT_PARSED="nil"
+	fi
+		
+	RUNCMD="${RUNCMD} $TEA_SERVER_URL_PARSED $TEA_SERVER_USERNAME_PARSED  $TEA_SERVER_PASSWORD_PARSED $PYTHONPATH $BE_INSTANCE_POLLAR_INTERVAL_PARSED  $CLUSTER_NAME_PARSED $AWS_EXECUTION_ENV $AWS_REGION_PARSED $AWS_OUTPUT_FORMAT_PARSED"
+  
+  elif [[ "$BE_INSTANCE_DISCOVERY_TYPE_PARSED" == "k8s" ]]; then #If discovery is for k8s
+	  BE_TEA_AGENT_URL_PARSED="${BE_TEA_AGENT_URL%}"
+      BE_TEA_AGENT_URL_PARSED="${BE_TEA_AGENT_URL_PARSED#\"}"
+      RUNCMD="${RUNCMD}  $TEA_SERVER_URL_PARSED $TEA_SERVER_USERNAME_PARSED  $TEA_SERVER_PASSWORD_PARSED $PYTHONPATH $BE_INSTANCE_POLLAR_INTERVAL_PARSED $BE_TEA_AGENT_URL_PARSED $BE_INSTANCE_DISCOVERY_TYPE_PARSED"
   fi
 else
 
 	if [ "$COMPONENT_PARSED" = "rms" ]; then
+  		TRA_FILE=$BE_HOME/rms/bin/be-rms.tra
 		mkdir -p $BE_HOME/eclipse-platform/eclipse/dropins && echo "path=/opt/tibco/be/$SHORT_VERSION/studio" > $BE_HOME/eclipse-platform/eclipse/dropins/TIBCOBusinessEvents-Studio-plugins.link
 	fi
